// src/app/modules/customer/core/models/customer.enum.ts
export enum EPartnerType {
  CORPORATE = "corporate",
  MAIN = "main",
  SUB = "sub",
  POINT_SALE = "point_sale",
}

export enum EPartnerTypeL {
  corporate = "Kurumsal",
  main = "Ana Bayi",
  sub = "Alt Bayi",
  point_sale = "Satış Noktasi",
}

export enum EPartnerScope {
  DASHBOARD = "dashboard",
  API = "api",
  COUPON = "coupon",
}

export enum EPartnerScopeL {
  dashboard = "Panel",
  api = "Api",
  coupon = "Kupon",
}

export enum ECustomerType {
  CORPORATE = "corporate",
  INDIVIDUAL = "individual",
}

export enum ECustomerTypeL {
  corporate = "Kurumsal",
  individual = "Bireysel",
}